+++ Sources/Core/Classes/models/requests/RefreshRequest.swift
@@ -3,13 +3,13 @@ import Foundation
 public class RefreshRequest: Codable, DictionaryEncodable {
     public let clientId: String
     public let refreshToken: String?
-    public let redirectUri: String
+    public let redirectUri: URL
     public let grantType: String
-    
+
     public init(
         clientId: String,
         refreshToken: String?,
-        redirectUri: String
+        redirectUri: URL
     ) {
         self.clientId = clientId
         self.refreshToken = refreshToken